Refactor/correctness

+++ src/main/java/de/hpi/msc/jschneider/fileHandling/writing/SequenceWriter.java
@@ -4,7 +4,7 @@ import de.hpi.msc.jschneider.utility.dataTransfer.DataSink;
 
 public interface SequenceWriter extends DataSink
 {
-    void write(float[] records);
+    void write(double[] records);
 
     boolean isNull();
