+++ src/main/java/de/hpi/msc/jschneider/protocol/sequenceSliceDistribution/SequenceSliceDistributionEvents.java
@@ -15,8 +15,8 @@ public class SequenceSliceDistributionEvents
         private static final long serialVersionUID = 3362960763906428742L;
         private long firstSubSequenceIndex;
         private boolean isLastSubSequenceChunk;
-        private float minimumRecord;
-        private float maximumRecord;
+        private double minimumRecord;
+        private double maximumRecord;
         private MatrixStore<Double> projection;
 
         @Override
